+++ programs/voter-stake-registry/tests/test_deposit_daily_vesting.rs
@@ -223,14 +223,6 @@ async fn test_deposit_daily_vesting() -> Result<(), TransportError> {
     assert_eq!(after_withdraw.vault, 6500);
     assert_eq!(after_withdraw.deposit, 6500);
 
-    // check that creating the voter again has no effect on balances
-    addin
-        .create_voter(&registrar, &token_owner_record, &voter_authority, &payer)
-        .await;
-
-    let after_recreate = get_balances(0).await;
-    assert_eq!(after_withdraw, after_recreate);
-
     // advance another day
     addin
         .set_time_offset(&registrar, &realm_authority, 73 * 60 * 60)
